fix: Fix mapview leaflet error
