+++ src/app/upload/SingleImage.js
@@ -1,3 +1,4 @@
+"use client";
 import { useState, useEffect } from "react";
 import MapView from "./MapView";
